workshops/taxi-pipeline/pyproject.toml
[project]
name = "taxi-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dlt[workspace]>=1.22.1",
    "marimo",
    "ibis-framework[duckdb]",
    "plotly",
]
